Moved utility function for link parsing to module_utils.
